YQL-19747 Extract NameSet
